malloc concat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -12,7 +12,7 @@
 
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
-	long double l1 = strlen(s1);
+	unsigned int l1 = strlen(s1);
 	unsigned int l2 = strlen(s2);
 	char *new_str;
